pyproject.toml
[project]
name = "gmail-cleanup"
version = "0.1.0"
description = "Safe Gmail cleanup CLI (dry-run first)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12",
  "rich>=13.7",
  "google-api-python-client>=2.140",
  "google-auth>=2.30",
  "google-auth-oauthlib>=1.2",
  "pyyaml>=6.0",
  "fastapi>=0.111",
  "uvicorn>=0.30",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "cryptography>=42.0",
  "python-dotenv>=1.0",
]

[project.scripts]
gmail-cleanup = "gmail_cleanup.cli:app"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["gmail_cleanup"]

